unlock/requirements.txt
Flask==2.3.2
google-cloud-storage==2.10.0
PyPDF2==3.0.1
python-keycloak==3.7.0
requests==2.31.0
python-jose==3.3.0
cryptography==41.0.7